+++ pyteal/ast/subroutine_test.py
@@ -1302,3 +1302,159 @@ def test_evaluate_subroutine_10_args():
         actual, _ = declaration.__teal__(options)
         options.setSubroutine(None)
         assert actual == expected
+
+
+def test_docstring_parsing_with_different_format():
+    short_desc = "Example of a ABIReturnSubroutine with short description docstring."
+    a_doc = "an abi Uint64 value"
+    return_doc = "A PyTeal expression that sets output Uint64 value as argument a."
+    long_desc = """Example first line.
+
+    This is a second line.
+
+    This is a third line that's so long it has to wrap in order to fit properly
+    in a line of source code.
+    """
+    expected_long_desc = "Example first line.\nThis is a second line.\nThis is a third line that's so long it has to wrap in order to fit properly in a line of source code."
+
+    def documented_method(a: pt.abi.Uint64, *, output: pt.abi.Uint64):
+        return output.set(a)
+
+    # Google format
+    documented_method.__doc__ = f"""{short_desc}
+
+    Args:
+        a: {a_doc}
+
+    Returns:
+        {return_doc}
+    """
+
+    mspec_dict = pt.ABIReturnSubroutine(documented_method).method_spec().dictify()
+    assert mspec_dict["desc"] == short_desc
+    assert mspec_dict["args"][0]["desc"] == a_doc
+    assert mspec_dict["returns"]["desc"] == return_doc
+
+    # epy format
+    documented_method.__doc__ = f"""
+    {short_desc}
+
+    @param a: {a_doc}
+    @return: {return_doc}
+    """
+
+    mspec_dict = ABIReturnSubroutine(documented_method).method_spec().dictify()
+    assert mspec_dict["desc"] == short_desc
+    assert mspec_dict["args"][0]["desc"] == a_doc
+    assert mspec_dict["returns"]["desc"] == return_doc
+
+    # numpy format
+    documented_method.__doc__ = f"""{short_desc}
+
+    Parameters
+    ----------
+    a:
+        an abi Uint64 value
+    output:
+        {a_doc}
+
+    Returns
+    -------
+    uint64
+        {return_doc}
+    """
+
+    mspec_dict = ABIReturnSubroutine(documented_method).method_spec().dictify()
+    assert mspec_dict["desc"] == short_desc
+    assert mspec_dict["args"][0]["desc"] == a_doc
+    assert mspec_dict["returns"]["desc"] == return_doc
+
+    # rst format
+    documented_method.__doc__ = f"""{short_desc}
+
+    :param a: {a_doc}
+    :returns: {return_doc}
+    """
+
+    mspec_dict = ABIReturnSubroutine(documented_method).method_spec().dictify()
+    assert mspec_dict["desc"] == short_desc
+    assert mspec_dict["args"][0]["desc"] == a_doc
+    assert mspec_dict["returns"]["desc"] == return_doc
+
+    # Short and long descriptions
+    documented_method.__doc__ = f"""{long_desc}
+
+    :param a: {a_doc}
+    :returns: {return_doc}
+    """
+
+    mspec_dict = ABIReturnSubroutine(documented_method).method_spec().dictify()
+    assert mspec_dict["desc"] == expected_long_desc
+    assert mspec_dict["args"][0]["desc"] == a_doc
+    assert mspec_dict["returns"]["desc"] == return_doc
+
+    # Only long description
+    # Short description is defined as being on the first line, so by introducing
+    # long_desc on the second, there is no short description.
+    documented_method.__doc__ = f"""
+    {long_desc}
+
+    :param a: {a_doc}
+    :returns: {return_doc}
+    """
+
+    mspec_dict = ABIReturnSubroutine(documented_method).method_spec().dictify()
+    assert mspec_dict["desc"] == expected_long_desc
+    assert mspec_dict["args"][0]["desc"] == a_doc
+    assert mspec_dict["returns"]["desc"] == return_doc
+
+    # No description
+    documented_method.__doc__ = f"""
+
+    :param a: {a_doc}
+    :returns: {return_doc}
+    """
+    mspec_dict = ABIReturnSubroutine(documented_method).method_spec().dictify()
+    assert "descr" not in mspec_dict
+    assert mspec_dict["args"][0]["desc"] == a_doc
+    assert mspec_dict["returns"]["desc"] == return_doc
+
+    algobank_example = """Withdraw an amount of Algos held by this app.
+
+    The sender of this method call will be the source of the Algos, and the destination will be
+    the `recipient` argument.
+
+    The Algos will be transferred to the recipient using an inner transaction whose fee is set
+    to 0, meaning the caller's transaction must include a surplus fee to cover the inner
+    transaction.
+
+    Args:
+        amount: The amount of Algos requested to be withdraw, in microAlgos. This method will fail
+            if this amount exceeds the amount of Algos held by this app for the method call sender.
+        recipient: An account who will receive the withdrawn Algos. This may or may not be the same
+            as the method call sender.
+    """
+
+    # algobank example
+    def withdraw(amount: pt.abi.Uint64, recipient: pt.abi.Account):
+        return pt.Assert(pt.Int(1))
+
+    withdraw.__doc__ = algobank_example
+
+    mspec_dict = ABIReturnSubroutine(withdraw).method_spec().dictify()
+    assert (
+        mspec_dict["desc"]
+        == "Withdraw an amount of Algos held by this app.\nThe sender of this method call will be the source of the Algos, "
+        + "and the destination will be the `recipient` argument.\nThe Algos will be transferred to the recipient using an inner transaction whose fee is "
+        + "set to 0, meaning the caller's transaction must include a surplus fee to cover the inner transaction."
+    )
+    assert (
+        mspec_dict["args"][0]["desc"]
+        == "The amount of Algos requested to be withdraw, in microAlgos. This method will fail if this amount exceeds "
+        + "the amount of Algos held by this app for the method call sender."
+    )
+    assert (
+        mspec_dict["args"][1]["desc"]
+        == "An account who will receive the withdrawn Algos. This may or may not be the same as the method call sender."
+    )
+    assert "desc" not in mspec_dict["returns"]